refactor: make internship as distinct app
