+++ backend/backend/settings.py
@@ -42,6 +42,7 @@ INSTALLED_APPS = [
     "corsheaders",
     'rest_framework.authtoken',
     'base',
+    'internship'
 ]
 
 MIDDLEWARE = [
