+++ backend/base/views.py
@@ -1,9 +1,9 @@
 from django.shortcuts import render
 from rest_framework.decorators import api_view
 from rest_framework.response import Response
-from .models import Internship, User, Student, InternshipLike
+from .models import User, Student
 from rest_framework import generics, permissions, status
-from .serializers import InternshipSerializer, UserSerializer, CompanySignupSerializer, StudentSignupSerializer, StudentSerializer, InternshipLikeSerializer
+from .serializers import UserSerializer, CompanySignupSerializer, StudentSignupSerializer, StudentSerializer
 from rest_framework.generics import (ListCreateAPIView,RetrieveUpdateDestroyAPIView, GenericAPIView, CreateAPIView) 
 from rest_framework.permissions import IsAuthenticated, IsAdminUser
 from .permissions import IsOwnerProfileOrReadOnly
@@ -11,40 +11,13 @@ from rest_framework.authtoken.models import Token
 from rest_framework.views import APIView
 from .permissions import IsStudentUser, IsCompanyUser
 from rest_framework.authtoken.views import ObtainAuthToken
-from django.shortcuts import get_object_or_404
 
-# Create your views here.
-
-@api_view(['GET'])
-def getInternships(request):
-    internships = Internship.objects.all()
-    serializer = InternshipSerializer(internships, many=True)
-    return Response(serializer.data)
 
 @api_view(['GET'])
 def getStudent(request, pk):
     student = Student.objects.get(user = pk)
     serializer = StudentSerializer(student, many=False)
     return Response(serializer.data)
-class InternshipListView(generics.ListCreateAPIView):
-    model = Internship
-    serializer_class = InternshipSerializer
-
-    def get_queryset(self):
-        queryset = Internship.objects.all()
-        name = self.request.query_params.get('name')
-        location = self.request.query_params.get('location')
-        category = self.request.query_params.get('category')
-        if name != None:
-            queryset = queryset.filter(name__icontains=name)
-            print(queryset)
-        elif location != None:
-            queryset = queryset.filter(location__icontains=location)
-            print(queryset)
-        elif category != None:
-            queryset = queryset.filter(category__icontains=category)
-            print(queryset)
-        return queryset
 
 class UserListCreateView(ListCreateAPIView):
     queryset= User.objects.all()
@@ -73,15 +46,6 @@ class CompanySignupView(GenericAPIView):
             "token":Token.objects.get(user=user).key,
             "message":"account created successfully"
         })
-
-def send_email(request):
-    n = random.randint(100000,999999)
-    email = EmailMessage(
-        'Please verify your email',
-        'This is your verification number' + str(n),
-        to=[UserSerializer.email]
-    )
-    email.send()
 
 
 class StudentSignupView(GenericAPIView):
@@ -130,33 +94,3 @@ class CompanyOnlyView(generics.RetrieveAPIView):
     def get_object(self):
         return self.request.user
 
-class InternshipLikeAPIView(generics.CreateAPIView):
-    """
-    Like, Dislike a recipe
-    """
-    serializer_class = InternshipLikeSerializer
-    permission_classes =[]
-
-    def post(self, request, pk):
-        print(request)
-        internship = get_object_or_404(Internship, _id=self.kwargs['pk'])
-        student = get_object_or_404(Student, user=request.user.id)
-        print(internship)
-        print(student)
-        new_like, created = InternshipLike.objects.get_or_create(
-            student=student, internship = internship)
-        if created:
-            new_like.save()
-            return Response(status=status.HTTP_201_CREATED)
-        return Response(status=status.HTTP_400_BAD_REQUEST)
-
-    def delete(self, request, pk):
-        internship = get_object_or_404(Internship, id=self.kwargs['pk'])
-        like = InternshipLike.objects.filter(student=request.user, internship=internship)
-        if like.exists():
-            like.delete()
-            return Response(status=status.HTTP_200_OK)
-        return Response(status=status.HTTP_400_BAD_REQUEST)
-
-    def perform_create(self, serializer):
-        serializer.save(author=self.request.user)
